requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
slowapi>=0.1.9
pytest>=8.0.0
pytest-asyncio>=0.24.0
